gdbm_fetch: data not always \0 - delimited
The data in a gdbm database can be arbitrary binary data and does not
have to be a zero-delimited C string. In fact some C applications store
C strings in a gdbm database without the trailing \0.

So in fetch(), when copying data from struct datum, copy datum.dsize
bytes. Then check if the last byte is \0 and if so, chop it off.