correctness update for 'datum'
In several places a 'datum' is constructed without checking the
length of the Rust string we're converting into a datum. So now
we check it, and if it's longer than i32::MAX we error out.

Similarly, we should also check the validity of a 'datum'
we get from the C size. Not only check if content.dptr.is_null,
but also content.dsize < 0 and return an error.

Because of this we have to change the signature of the
'store' method to Result<bool, GdbmError> instead of -1/0/1.
But that is correct anyway, as -1 indicates an error.

Update test test  as well.